requirements.txt
geopandas==0.10.2
keras==2.8.0
Keras-Preprocessing==1.1.2
matplotlib==3.2.2
matplotlib-inline==0.1.3
numpy==1.21.6
opencv-python==4.1.2.30
pandas==1.3.5
Pillow==7.1.2
rasterio==1.2.10
sklearn==0.0
sklearn-pandas==1.8.0
tensorflow==2.8.0
